+++ cmd/binder/app/app.go
@@ -150,6 +150,9 @@ func (app *App) RegisterPlugins(plugins *plugins.BinderPlugins) {
 	app.plugins = plugins
 }
 
+// +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=coordination.k8s.io,resources=leases,verbs=get;list;watch;create;update;patch;delete
+
 func (app *App) Run() error {
 	var err error
 	go func() {