+++ cmd/nodescaleadjuster/app/app.go
@@ -33,6 +33,10 @@ func init() {
 	// +kubebuilder:scaffold:scheme
 }
 
+// +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=coordination.k8s.io,resources=leases,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=core,resources=events,verbs=create;patch;update
+
 func Run() error {
 	options := NewOptions()
 	options.AddFlags()
@@ -49,7 +53,9 @@ func Run() error {
 
 	clientConfig := ctrl.GetConfigOrDie()
 	mgr, err := ctrl.NewManager(clientConfig, ctrl.Options{
-		Scheme: scheme,
+		Scheme:           scheme,
+		LeaderElection:   options.EnableLeaderElection,
+		LeaderElectionID: "7n7h49uc.kai.scheduler",
 	})
 	if err != nil {
 		setupLog.Error(err, "unable to start manager")