+++ cmd/nodescaleadjuster/app/options.go
@@ -17,6 +17,7 @@ type Options struct {
 	GPUMemoryToFractionRatio float64
 	ScalingPodAppLabel       string
 	ScalingPodServiceAccount string
+	EnableLeaderElection     bool
 }
 
 // NewOptions creates a new Options
@@ -26,23 +27,27 @@ func NewOptions() *Options {
 }
 
 // AddFlags adds flags for a specific CMServer to the specified FlagSet
-func (s *Options) AddFlags() {
-	flag.StringVar(&s.ScalingPodImage,
+func (o *Options) AddFlags() {
+	flag.StringVar(&o.ScalingPodImage,
 		"scaling-pod-image", consts.DefaultScalingPodImage,
 		"The image to use for the scaling pod, defaults to "+consts.DefaultScalingPodImage)
-	flag.StringVar(&s.ScalingPodNamespace,
+	flag.StringVar(&o.ScalingPodNamespace,
 		"scale-adjust-namespace", "kai-scale-adjust",
 		"The namespace to use for the scaling pods, defaults to kai-scale-adjust")
-	flag.StringVar(&s.SchedulerName,
+	flag.StringVar(&o.SchedulerName,
 		"scheduler-name", "kai-scheduler",
 		"Scheduler name, defaults to kai-scheduler")
-	flag.StringVar(&s.ScalingPodAppLabel,
+	flag.StringVar(&o.ScalingPodAppLabel,
 		"scaling-pod-app-label", "scaling-pod",
 		"Scaling pod app label")
-	flag.StringVar(&s.ScalingPodServiceAccount,
+	flag.StringVar(&o.ScalingPodServiceAccount,
 		"scaling-pod-service-account", "scaling-pod",
 		"Scaling pod service account name")
-	flag.Float64Var(&s.GPUMemoryToFractionRatio,
+	flag.Float64Var(&o.GPUMemoryToFractionRatio,
 		"gpu-memory-to-fraction-ratio", consts.DefaultGPUMemoryToFractionRatio,
 		fmt.Sprintf("The ratio of GPU memory to fraction, defaults to %f", consts.DefaultGPUMemoryToFractionRatio))
+	flag.BoolVar(&o.EnableLeaderElection,
+		"leader-elect", false,
+		"Enable leader election for controller manager. "+
+			"Enabling this will ensure there is only one active controller manager.")
 }