+++ cmd/podgrouper/app/app.go
@@ -47,6 +47,9 @@ func init() {
 	// +kubebuilder:scaffold:scheme
 }
 
+// +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=coordination.k8s.io,resources=leases,verbs=get;list;watch;create;update;patch;delete
+
 func Run() error {
 	var opts Options
 	opts.AddFlags(flag.CommandLine)